# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "ActionRequest": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "notes": {
            "default": "",
            "title": "Notes",
            "type": "string"
          }
        },
        "required": [
          "action"
        ],
        "title": "ActionRequest",
        "type": "object"
      },
      "ActionResponse": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "job_id": {
            "title": "Job Id",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "job_id",
          "action"
        ],
        "title": "ActionResponse",
        "type": "object"
      },
      "ActionsListResponse": {
        "properties": {
          "actions": {
            "items": {
              "$ref": "#/components/schemas/ActionResponse"
            },
            "title": "Actions",
            "type": "array"
          }
        },
        "required": [
          "actions"
        ],
        "title": "ActionsListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunEntry": {
        "description": "Single ``run_log`` row exposed over the API.\n\nMaps every column from the baseline + migration-0010 schema.  Optional\nfields may be NULL on rows inserted before the observability migration\nran; Pydantic coerces them to None rather than raising.",
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "new_jobs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Jobs"
          },
          "per_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Source"
          },
          "per_source_duration": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Source Duration"
          },
          "per_source_errors": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Per Source Errors"
          },
          "run_uuid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Uuid"
          },
          "sources_queried": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sources Queried"
          },
          "timestamp": {
            "title": "Timestamp",
            "type": "string"
          },
          "total_duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Duration"
          },
          "total_found": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Found"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id"
          }
        },
        "required": [
          "id",
          "timestamp"
        ],
        "title": "RunEntry",
        "type": "object"
      },
      "RunsListResponse": {
        "properties": {
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunEntry"
            },
            "title": "Runs",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "runs",
          "total",
          "limit",
          "offset"
        ],
        "title": "RunsListResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Job360 API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/actions": {
      "get": {
        "operationId": "list_actions_api_actions_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "job360_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job360 Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Actions",
        "tags": [
          "actions"
        ]
      }
    },
    "/api/actions/counts": {
      "get": {
        "operationId": "action_counts_api_actions_counts_get",
        "parameters": [
          {
            "in": "cookie",
            "name": "job360_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job360 Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "integer"
                  },
                  "title": "Response Action Counts Api Actions Counts Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Action Counts",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/api/jobs/{job_id}/action": {
      "delete": {
        "operationId": "delete_action_api_jobs__job_id__action_delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "job360_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job360 Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Action",
        "tags": [
          "actions"
        ]
      },
      "post": {
        "operationId": "set_action_api_jobs__job_id__action_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "job360_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job360 Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Action",
        "tags": [
          "actions"
        ]
      }
    },
...
    "/api/runs/recent": {
      "get": {
        "description": "Return paginated pipeline run history from run_log, newest first.\n\nStep-3 B-15.  Requires authentication (rule #12) \u2014 the run log is\noperational metadata, not a public catalog surface.",
        "operationId": "recent_runs_api_runs_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "job360_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Job360 Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}